linux: Add support for clusters if provided by sysfs.
These allow us to represent groups of cores where they are grouped
because of shared elements other than caches.  Typical example
would be a heirarchical coherency fabric with a cluster sharing a
single bridge to a wider coherency network on chip.

In ACPI this is represented in PPTT as a processor hieracrhy node
structure above the level representing a CPU (possibly with multiple
threads) and below the hierarchy node structure that represents
either a die, or the level at which the last level cache is shared.
An example of this is the Kunpeng920 SoC which has 4 CPU cores
per cluster and up to 8 of those clusters share an L3 cache.

Similar concepts exist in device tree.

I am not sure what value is appropriate for obj_type_property[].
Currently I have set it to just below that of the die on the
basis that was an older concept and hence we should use that
in preference to the cluster if they happen to match.

Kernel patch exporting the sysfs ABI will be sent to the
relevant mailing lists shortly.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>